third_party/WebKit/LayoutTests/media/no-autoplay-with-user-gesture-requirement.html

<!DOCTYPE html>
<title>Test that media autoplay should not work if user gesture is required for playback.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
internals.settings.setAutoplayPolicy('user-gesture-required');
</script>
<video autoplay></video>
<script>
async_test(function(t) {
    var video = document.querySelector("video");
    video.src = "content/test.ogv";
    assert_true(video.paused);

    video.onplay = t.unreached_func();

    video.oncanplaythrough = t.step_func(function() {
        setTimeout(t.step_func_done(function() { assert_true(video.paused); }), 100);
    });
});
</script>
